requirements.txt
numpy
librosa
sounddevice
scipy
Pillow
